story #524 School Funding
git-svn-id: http://192.168.110.32/svn/Products/PFF/Base/Branches/MLOD@94180 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennanttech/pff/incomeexpensedetail/dao/IncomeExpenseDetailDAO.java
@@ -19,6 +19,6 @@ public interface IncomeExpenseDetailDAO {
 	
 	List<IncomeExpenseDetail> getExpenseList(Long id, String incomeType, String type);
 	
-	List<Map<String, Object>> getTotal(Long custId, Long financialYear);
+	List<Map<String, Object>> getTotal(Long custId, Integer financialYear);
 
 }
